Fix clippy warnings

+++ src/editor.rs
@@ -644,7 +644,7 @@ impl Editor {
             current = (current + if forward { 1 } else { num_rows - 1 }) % num_rows;
             let row = &mut self.rows[current];
             if let Some(cx) = slice_find(&row.chars, query.as_bytes()) {
-                self.cursor.y = current as usize;
+                self.cursor.y = current;
                 self.cursor.x = cx;
                 // Try to reset the column offset; if the match is after the offset, this
                 // will be updated in self.cursor.scroll() so that the result is visible
